+++ app/src/androidTest/java/woowacourse/payments/ui/components/PaymentCardPlateTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.features.cartinput
+package woowacourse.payments.ui.features.cardinput
 
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.graphics.Color
