+++ app/src/androidTest/java/woowacourse/payments/ui/features/cardinput/components/CardExpireDateFieldTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.features.cartinput.components
+package woowacourse.payments.ui.features.cardinput.components
 
 import androidx.activity.ComponentActivity
 import androidx.compose.runtime.getValue
