+++ app/src/main/java/woowacourse/payments/data/PaymentCardRecord.kt
@@ -1,6 +1,6 @@
 package woowacourse.payments.data
 
-import woowacourse.payments.ui.features.cartinput.CardUiState
+import woowacourse.payments.ui.features.cardinput.CardUiState
 
 data class PaymentCardRecord(
     val id: Int = -1,
