+++ app/src/main/java/woowacourse/payments/data/PaymentInMemoryRepository.kt
@@ -1,6 +1,6 @@
 package woowacourse.payments.data
 
-import woowacourse.payments.ui.features.cartinput.CardUiState
+import woowacourse.payments.ui.features.cardinput.CardUiState
 import java.util.concurrent.atomic.AtomicInteger
 
 object PaymentInMemoryRepository {
