+++ app/src/main/java/woowacourse/payments/ui/features/cardinput/components/CardPasswordField.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.features.cartinput.components
+package woowacourse.payments.ui.features.cardinput.components
 
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.ExperimentalMaterial3Api
